basic-pitch-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
basic-pitch==0.2.4
numpy==1.24.3
librosa==0.10.0
scipy==1.11.1
soundfile==0.12.1
python-dotenv==1.0.0
gunicorn==21.2.0

